smart bgm and hursty themes

+++ scriptmodules/config/bgmtoggle.sh
@@ -23,9 +23,13 @@ function gui_bgmtoggle() {
     if [[ -n "$choice" ]]; then
         case "$choice" in
             1)
-                pkill -CONT mpg123
+                if [[ ! -f "$HOME/.config/bgmtoggle" ]]; then
+                    (mpg123 -f 18000 -Z /home/pigaming/bgm/*.mp3 >/dev/null 2>&1)
+                else
+                    pkill -CONT mpg123
+                fi
                 touch "$HOME/.config/bgmtoggle"
-                printMsgs "dialog" "Background music is now enabled.\n\nA reboot may be required for take in effect."
+                printMsgs "dialog" "Background music is now enabled."
                 ;;
             2)
                 pkill -STOP mpg123
